# docs/swagger.yaml
definitions:
  schema.SaveSongInput:
    properties:
      group:
        type: string
      song:
        type: string
    type: object
  schema.Song:
    properties:
      group:
        type: string
      id:
        type: string
      link:
        type: string
      releasedate:
        type: string
      song:
        type: string
      text:
        type: string
    required:
    - group
    - link
    - releasedate
    - song
    - text
    type: object
  schema.SongOutput:
    properties:
      group:
        type: string
      link:
        type: string
      releasedate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
paths:
  /songs:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: group
        type: string
      - in: query
        name: link
        type: string
      - in: query
        name: releasedate
        type: string
      - in: query
        name: song
        type: string
      - in: query
        name: text
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/schema.Song'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all songs
      tags:
      - songs
  /songs/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Song id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.Song'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retrieve one song
      tags:
      - songs
    post:
      consumes:
      - application/json
      parameters:
      - description: Input data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/schema.SaveSongInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.SongOutput'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Save song
      tags:
      - songs
  /songs/{id}/text:
    get:
      consumes:
      - application/json
      parameters:
      - description: Song id
        in: path
        name: id
        required: true
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                type: string
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get song text
      tags:
      - songs
swagger: "2.0"